+++ tests/testthat/test-checkNoSubfolder.R
@@ -1,3 +1,0 @@
-test_that("Verify there are no subfolders", {
-  expect_true(raw.checkNoSubfolders(raw.getSamplePath()))
-})